# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loadout HTTP API",
    "version": "0.2.0",
    "description": "实际Go app/identity/运维路由。网页会话与ldt_网关令牌分离；业务写请求校验Origin；外部支付目前不可用。Body总大小app默认16KiB、技能发布50MiB（另验32文件/单文件256KiB）、identity4KiB；字符串长度限制以Go UTF-8字节校验为准，JSON Schema maxLength仅作客户端提示。"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源服务根地址"
    }
  ],
...
  "security": [
    {
      "CookieSession": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "post_api_v1_auth_register",
        "summary": "注册并创建会话",
        "tags": ["账号"],
        "description": "注册并创建会话",
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResult"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "loadout_session AT 与 loadout_refresh RT；均 HttpOnly、SameSite=Lax、Path=/，生产Secure；默认900/604800秒，系统配置可热更新。"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "format": "uri"
            },
            "description": "必须与部署公开站点源一致；浏览器同源写请求自动携带。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "post_api_v1_auth_login",
        "summary": "登录并创建会话",
        "tags": ["账号"],
        "description": "登录并创建会话",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResult"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "loadout_session AT 与 loadout_refresh RT；均 HttpOnly、SameSite=Lax、Path=/，生产Secure；默认900/604800秒，系统配置可热更新。"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "format": "uri"
            },
            "description": "必须与部署公开站点源一致；浏览器同源写请求自动携带。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "post_api_v1_auth_logout",
        "summary": "注销当前会话",
        "tags": ["账号"],
        "description": "删除当前Cookie对应会话并清Cookie；没有Cookie也返回204。",
        "responses": {
          "204": {
            "description": "成功"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "format": "uri"
            },
            "description": "必须与部署公开站点源一致；浏览器同源写请求自动携带。"
          }
        ]
      }
    },
    "/api/v1/account/me": {
      "get": {
        "operationId": "get_api_v1_account_me",
        "summary": "当前账号与用量概览",
        "tags": ["账号"],
        "description": "今日与本月边界按UTC；实际余额来自钱包，注册赠额可由部署配置修改。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ]
      }
    },
    "/api/v1/account/verify": {
      "get": {
        "operationId": "get_api_v1_account_verify",
        "summary": "验证网关令牌",
        "tags": ["账号"],
        "description": "仅ldt_ Bearer，返回该令牌钱包余额与启用预设池key。远程工具的具体命名空间工具用MCP tools/list查询。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenVerify"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "GatewayToken": []
          }
        ]
      }
    },
    "/api/v1/account/tokens": {
      "get": {
        "operationId": "get_api_v1_account_tokens",
        "summary": "个人创建的令牌列表",
        "tags": ["账号"],
        "description": "个人创建的令牌列表",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ]
      },
      "post": {
        "operationId": "post_api_v1_account_tokens",
        "summary": "创建个人或团队令牌",
        "tags": ["账号"],
        "description": "创建个人或团队令牌",
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenCreated"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "format": "uri"
            },
            "description": "必须与部署公开站点源一致；浏览器同源写请求自动携带。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenCreate"
              }
            }
          }
        }
      }
    },
    "/api/v1/account/tokens/{id}": {
      "delete": {
        "operationId": "delete_api_v1_account_tokens_id",
        "summary": "撤销自己的令牌",
        "tags": ["账号"],
        "description": "重复撤销保持原撤销时间；其他账号令牌不可操作。",
        "responses": {
          "204": {
            "description": "成功"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "format": "uri"
            },
            "description": "必须与部署公开站点源一致；浏览器同源写请求自动携带。"
          }
        ]
      }
    },
    "/api/v1/account/usage": {
      "get": {
        "operationId": "get_api_v1_account_usage",
        "summary": "个人调用明细",
        "tags": ["账号"],
        "description": "ID降序游标；团队接口要求当前成员。status=ok代表成功，error代表失败。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsagePage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/ToolFilter"
          },
          {
            "$ref": "#/components/parameters/StatusFilter"
          }
        ]
      }
    },
    "/api/v1/admin/usage": {
      "get": {
        "operationId": "get_api_v1_admin_usage",
        "summary": "全局调用明细",
        "tags": ["管理"],
        "description": "ID降序游标；团队接口要求当前成员。status=ok代表成功，error代表失败。 仅管理员，角色由服务端验证。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsagePage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/ToolFilter"
          },
          {
            "$ref": "#/components/parameters/StatusFilter"
          },
          {
            "$ref": "#/components/parameters/UserFilter"
          }
        ]
      }
    },
    "/api/v1/admin/usage/export": {
      "get": {
        "operationId": "get_api_v1_admin_usage_export",
        "summary": "分页导出调用CSV",
        "tags": ["管理"],
        "description": "分页导出调用CSV 仅管理员，角色由服务端验证。",
        "responses": {
          "200": {
            "description": "单页CSV，字符串单元格进行公式转义。",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "headers": {
              "X-Next-Cursor": {
                "schema": {
                  "type": "string"
                },
                "description": "下一页cursor，空字符串表示结束。"
              },
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "example": "attachment; filename=\"loadout-usage.csv\""
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/ToolFilter"
          },
          {
            "$ref": "#/components/parameters/StatusFilter"
          },
          {
            "$ref": "#/components/parameters/UserFilter"
          }
        ]
      }
    },
    "/api/v1/admin/usage/summary": {
      "get": {
        "operationId": "get_api_v1_admin_usage_summary",
        "summary": "全局工具汇总",
        "tags": ["管理"],
        "description": "UTC今天或含今天的最近7天；按分组游标分页，默认50/最大100组；消费方须继续读取next_cursor直到为空；仅按days汇总，明细筛选不改变此汇总。 仅管理员，角色由服务端验证。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ToolSummary"
                      }
                    },
                    "next_cursor": {
                      "type": "string"
                    }
                  },
                  "required": ["items", "next_cursor"]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Days"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ]
      }
    },
    "/api/v1/account/teams/{id}/usage": {
      "get": {
        "operationId": "get_api_v1_account_teams_id_usage",
        "summary": "团队调用明细",
        "tags": ["账号"],
        "description": "ID降序游标；团队接口要求当前成员。status=ok代表成功，error代表失败。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsagePage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/ToolFilter"
          },
          {
            "$ref": "#/components/parameters/StatusFilter"
          },
          {
            "$ref": "#/components/parameters/UserFilter"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ]
      }
    },
    "/api/v1/account/teams/{id}/usage/export": {
      "get": {
        "operationId": "get_api_v1_account_teams_id_usage_export",
        "summary": "分页导出调用CSV",
        "tags": ["账号"],
        "description": "分页导出调用CSV",
        "responses": {
          "200": {
            "description": "单页CSV，字符串单元格进行公式转义。",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "headers": {
              "X-Next-Cursor": {
                "schema": {
                  "type": "string"
                },
                "description": "下一页cursor，空字符串表示结束。"
              },
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "example": "attachment; filename=\"loadout-usage.csv\""
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/ToolFilter"
          },
          {
            "$ref": "#/components/parameters/StatusFilter"
          },
          {
            "$ref": "#/components/parameters/UserFilter"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ]
      }
    },
    "/api/v1/account/teams/{id}/usage/summary": {
      "get": {
        "operationId": "get_api_v1_account_teams_id_usage_summary",
        "summary": "团队成员汇总",
        "tags": ["账号"],
        "description": "UTC今天或含今天的最近7天；按分组游标分页，默认50/最大100组；消费方须继续读取next_cursor直到为空；仅按days汇总，明细筛选不改变此汇总。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MemberSummary"
                      }
                    },
                    "next_cursor": {
                      "type": "string"
                    }
                  },
                  "required": ["items", "next_cursor"]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Days"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "operationId": "get_api_v1_admin_users",
        "summary": "用户列表",
        "tags": ["管理"],
        "description": "用户列表 仅管理员，角色由服务端验证。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ]
      }
    },
    "/api/v1/admin/users/{id}": {
      "patch": {
        "operationId": "patch_api_v1_admin_users_id",
        "summary": "启用或停用用户",
        "tags": ["管理"],
        "description": "禁止停用自己或最后一个启用管理员。 仅管理员，角色由服务端验证。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "format": "uri"
            },
            "description": "必须与部署公开站点源一致；浏览器同源写请求自动携带。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": {
                    "type": "boolean"
                  }
                },
                "required": ["enabled"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plans": {
      "get": {
        "operationId": "get_api_v1_plans",
        "summary": "启用套餐列表",
        "tags": ["账号"],
        "description": "启用套餐列表",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "loadout_session",
        "description": "HttpOnly、SameSite=Lax，默认900秒opaque AT；生产Secure。旧7天session在原期限内兼容。"
      },
      "GatewayToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ldt_ token",
        "description": "网页会话不能替代此令牌。"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "loadout_refresh"
      }
    },
    "schemas": {
...
      "UserResult": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": ["user"]
      },
...
      "Credentials": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 32,
            "pattern": "^[a-zA-Z0-9_-]+$"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "description": "注册要求 UTF-8 长度 12–1024 字节；不是字符计数。",
            "writeOnly": true
          }
        },
        "required": ["username", "password"],
        "additionalProperties": false
      },
...
    },
...
    "responses": {
      "Error": {
        "description": "业务错误；参见 error 稳定代码。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}